feat(signals): relocate signals subtree + gatherer into the package
