+++ packages/signals/src/providers/coverage-trend.ts
