+++ packages/signals/src/registry.ts
