+++ packages/signals/src/shared.ts
