+++ packages/signals/src/types.ts
